+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project changes will be shown here.
 - Issue 550 - Rename 2nd Site links
 - Issue 547 - Time Request Denied Process
 - Issue 532 - Secondary LDAP server
+- Issue 529 - P5 default times when generated from Sales Order
 ### Fixed
 - Issue 551 - Contact Audit Log missing support level
 - Issue 552 - Contact Position field extension